testing edits

+++ SeasonManager.js
@@ -60,7 +60,7 @@ const TestingList = ({ navigation }) => {
 
       let changeButton = (item) =>
       {
-        if(currPlayer.email.localeCompare("admin") === 0) {
+        if((currPlayer) && currPlayer.email.localeCompare("admin") === 0) {
           return (
             <TouchableOpacity 
             onPress={() => deleteSeason(item)} >
